feat: `thenReject` now lazily creates promise rejection
Previously, `thenReject` would instantly create a `Promise.reject()`
that could cause problems with code that did not immediately call the
instance and capture the promise.

Now, the expectation sets an internal value that will be turned into a
promise rejection only when the instance is called. This should avoid
the unhandled rejection warnings that you can get when either 1) the
expectation is not met or 2) there is code that flushes promises before
they're actually caught, such as `act` from `@testing-library/react`.

Moreover, this should help avoid pausing a debugger on `thenReject` when
"Pause on any exception" is checked.

Fixes #238.